Remove /explore and rename API
