+++ .env.sample
@@ -1,7 +1,6 @@
 NEXT_PUBLIC_API_GEO_URL=https://geo.api.gouv.fr
 NEXT_PUBLIC_BACKEND_URL=https://backend.adresse.data.gouv.fr
-NEXT_PUBLIC_API_EXPLORE_URL=https://sandbox.geo.api.gouv.fr/explore
-NEXT_PUBLIC_API_EXPLORE_SEARCH_URL=https://api-adresse.data.gouv.fr
+NEXT_PUBLIC_API_BAN_URL=https://sandbox.geo.api.gouv.fr/explore
 NEXT_PUBLIC_API_ADRESSE=https://api-adresse.data.gouv.fr
 NEXT_PUBLIC_API_ETABLISSEMENTS_PUBLIC=https://etablissements-publics.api.gouv.fr/v3
 NEXT_PUBLIC_BAN_DASHBOARD_URL=http://build-1.infra.geo.data.gouv.fr/
