+++ components/hero.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import {Download, Edit3, Database} from 'react-feather'
 import theme from '@/styles/theme'
 import ToolsIcon from './icons/tools'
-import ExploreSearch from './explorer/explore-search'
+import BanSearch from './ban-search'
 
 import Container from './container'
 
@@ -54,7 +54,7 @@ function Hero({title, tagline}) {
             </Link>
           </div>
 
-          <ExploreSearch />
+          <BanSearch />
         </Container>
 
       </div>
