+++ lib/api-adresse.js
@@ -24,7 +24,7 @@ async function _fetch(url) {
 
 export function search(args) {
   const {q, lng, lat} = args
-  let url = `${API_ADRESSE}/search/?q=${q}`
+  let url = `${API_ADRESSE}/search/?q=${encodeURIComponent(q)}`
 
   if (lng && lat) {
     url += `&lng=${lng}&lat=${lat}`
