+++ lib/api-ban.js
@@ -1,7 +1,6 @@
-import HttpError from '../http-error'
+import HttpError from './http-error'
 
-export const API_EXPLORE_URL = process.env.NEXT_PUBLIC_API_EXPLORE_URL || 'https://sandbox.geo.api.gouv.fr/explore'
-export const API_EXPLORE_SEARCH_URL = process.env.NEXT_PUBLIC_API_EXPLORE_SEARCH_URL || 'https://api-adresse.data.gouv.fr'
+export const API_BAN_URL = process.env.NEXT_PUBLIC_API_BAN_URL || 'https://sandbox.geo.api.gouv.fr/explore'
 
 export async function _fetch(url) {
   const options = {
@@ -23,10 +22,6 @@ export async function _fetch(url) {
   throw new Error('Une erreur est survenue')
 }
 
-export function search(input) {
-  return _fetch(`${API_EXPLORE_SEARCH_URL}/search?q=${encodeURIComponent(input)}`)
-}
-
 export function getAddress(idAddress) {
-  return _fetch(`${API_EXPLORE_URL}/${idAddress}`)
+  return _fetch(`${API_BAN_URL}/${idAddress}`)
 }
